se refactoriza para no repetir tanto codigo
